Extract `ApplyCmd` from `EnsureCmd`.
