+++ crate/rt/src/cmds/sub.rs
@@ -6,11 +6,12 @@
 //! [`OutputWrite`]: peace_rt_model_core::OutputWrite
 
 pub use self::{
-    states_current_discover_cmd::StatesCurrentDiscoverCmd,
+    apply_cmd::ApplyCmd, states_current_discover_cmd::StatesCurrentDiscoverCmd,
     states_desired_discover_cmd::StatesDesiredDiscoverCmd,
     states_desired_read_cmd::StatesDesiredReadCmd, states_saved_read_cmd::StatesSavedReadCmd,
 };
 
+mod apply_cmd;
 mod states_current_discover_cmd;
 mod states_desired_discover_cmd;
 mod states_desired_read_cmd;
